Speed up the time-scons.py "build" of SCons with compiled *.pyc files, by just invoking setup.py directly instead of packing and unpacking the distributable .tar.gz file.
Refactor to be able to capture historical statistics (of previous
revisions) by preparing the "built" revision once and then running all
of the requested timing scripts, with output going to a separate log
file for each revision + script.

Fix TestSCons.py so that it interprets the scons script name to be tested
relative to the invoking directory, not always relative to src/script.