Fix for MSSDK initialiation
For the user in question who had VS 8.0 express installed, the vc version was being set to 8.0, and the the vs version was being set to the same, which was then being used to detect the vs version by an expected binary.
Since it though it had VS 8.0, the binary was the not there for VS8.0 Express, so it was bailing out on the initialization.

The fix is to have a VC8.0Exp as well as VC8.0 (also for VS9.0Exp).